Karate Man started.

+++ Assets/Scripts/Games/Spaceball/Spaceball.cs
@@ -34,7 +34,9 @@ namespace RhythmHeavenMania.Games.Spaceball
             for (int i = 1; i < BallsHolder.transform.childCount; i++)
                 Destroy(BallsHolder.transform.GetChild(i).gameObject);
             GameManager.instance.GameCamera.orthographic = false;
-            SpaceballPlayer.instance.EligibleHits.RemoveRange(0, SpaceballPlayer.instance.EligibleHits.Count);
+            if (SpaceballPlayer.instance != null)
+            if (SpaceballPlayer.instance.EligibleHits.Count > 0)
+                SpaceballPlayer.instance.EligibleHits.RemoveRange(0, SpaceballPlayer.instance.EligibleHits.Count);
         }
 
         public override void OnTimeChange()
